initial development

+++ run.sh
@@ -42,7 +42,13 @@ else
 
     if [[ ! -f $config ]]; then
 
-      cp httpd-minimal.conf $config
+      if [[ -f httpd-minimal.conf ]]; then
+
+        cp httpd-minimal.conf $config
+      elif [[ -f vendor/bravedave/esse/httpd-minimal.conf ]]; then
+
+        cp vendor/bravedave/esse/ $config
+      fi
 
       echo "CustomLog $access_log common" >>$config
       echo "DocumentRoot `pwd`/www" >>$config
